long unsigned

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,17 +7,17 @@
   */
 int main(void)
 {
-	int num1 = 1;
-	int num2 = 2;
-	int sum;
+	unsigned long int num1 = 1;
+	unsigned long int num2 = 2;
+	unsigned long int  sum;
 	int i;
 
-	printf("%d, %d", num1, num2);
+	printf("%lu, %lu", num1, num2);
 
 	for (i = 2; i < 96; i++)
 	{
 		sum = num1 + num2;
-		printf(", %d", sum);
+		printf(", %lu", sum);
 
 		num1 = num2;
 		num2 = sum;
